% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font-setup.R
\name{rand_font}
\alias{rand_font}
\title{Get the active randplot font family}
\usage{
rand_font()
}
\value{
A character string: the active font family name.
}
\description{
Returns the font family name that \code{theme_rand()} will use by default. This
is \code{"ABCMonumentGrotesk"} after a successful \code{rand_font_setup()}, or
\code{"sans"} if no font has been configured.
}
\seealso{
\code{\link[=rand_font_setup]{rand_font_setup()}}
}
