% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand-pal.R
\docType{data}
\name{rand_pal}
\alias{rand_pal}
\alias{RandCatPal}
\alias{RandGrayPal}
\alias{show_rand_pal}
\title{RAND color palettes for data visualization}
\format{
An object of class \code{character} of length 9.

An object of class \code{character} of length 9.
}
\usage{
RandCatPal

RandGrayPal

show_rand_pal(palette_name)
}
\arguments{
\item{palette_name}{One of "RandCatPal" or "RandGrayPal"}
}
\description{
Access colors from the RAND color palettes.
}
\details{
\code{RandCatPal} is a palette of nine colorblind-safe colors for discrete data. Some of the colors in the \code{RandCatPal} palette are similar to people with some kinds of colorblindness, but the colors are still distinguishable.

\code{RandGrayPal} is a palette of nine colorblind-safe shades of gray. It is provided for easier access to a standardized palette of grays, so it isn't necessarily well-suited for visualization.

\code{show_rand_pal()} accepts one of "RandCatPal" or "RandGrayPal" and shows its colors.
}
\section{Palettes}{

Use \code{RandCatPal} or \code{RandGrayPal} in a plot by passing one of them to \code{ggplot2::scale_color_manual()} or \code{ggplot2::scale_fill_manual()}.
\describe{
\item{RandCatPal}{
\itemize{
\item #61c6c6
\item #ce8644
\item #c1536b
\item #595a5a
\item #e2c755
\item #597cbe
\item #45aF84
\item #7995a5
\item #af61a7
}
}
\item{RandGrayPal}{
\itemize{
\item #f1f1f1
\item #d7d7d7
\item #bdbdbd
\item #a4a4a4
\item #8c8c8c
\item #747474
\item #5e5e5e
\item #484848
\item #333333
}
}
}
}

\examples{
\dontrun{
# Use RandCatPal in a plot ----------------------------------------
library(tidyverse)
library(scales)
library(gapminder)
library(randplot)

gdp <-
  gapminder \%>\%
  filter(continent == "Oceania")

ggplot(gdp, aes(x = year,
                y = gdpPercap,
                color = country)) +
  geom_line() +
  scale_x_continuous(expand = c(0, 0)) +
  scale_y_continuous(
    expand = c(0, 0),
    labels = dollar_format(),
    limits = c(0, max(gdp$gdpPercap) + 10000),
    position = "right"
  ) +
  scale_color_manual(values = RandCatPal) +
  xlab(NULL) +
  ylab("GDP per capita") +
  theme_rand()


# Preview palette colors ------------------------------------------

# RandCatPal colors
show_rand_pal("RandCatPal")

# RandGrayPal colors
show_rand_pal("RandGrayPal")
}
}
\keyword{datasets}
