% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-rand.R
\name{theme_rand}
\alias{theme_rand}
\title{Apply the RAND theme}
\usage{
theme_rand(font = "Helvetica")
}
\arguments{
\item{font}{"Helvetica" by default. Helvetica is used for graphics on rand.org. Pass "Helvetica Neue" for graphics that will be printed in a RAND report.}
}
\description{
This is the RAND ggplot2 theme. It controls the display of ggplot2 charts.
}
\details{
Use \code{theme_rand()} like other ggplot2 themes by adding \code{theme_rand()} to your plot.

\code{theme_rand()} uses Helvetica as its default font, which is used for graphics on rand.org. Graphics in printed RAND reports use Helvetica Neue, so use \code{theme_rand("Helvetica Neue")} if your graphics will go into a printed report.
}
\examples{
\dontrun{
library(tidyverse)
library(scales)
library(gapminder)
library(randplot)

gdp <-
  gapminder \%>\%
  filter(continent == "Oceania")

ggplot(gdp, aes(x = year,
                y = gdpPercap,
                color = country)) +
  geom_line() +
  scale_x_continuous(expand = c(0, 0)) +
  scale_y_continuous(
    expand = c(0, 0),
    labels = dollar_format(),
    limits = c(0, max(gdp$gdpPercap) + 10000),
    position = "right"
  ) +
  scale_color_manual(values = RandCatPal) +
  xlab(NULL) +
  ylab("GDP per capita") +
  theme_rand()
}
}
